+++ app/src/main/java/com/codesoom/assignment/services/UserRegisterService.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.application;
+package com.codesoom.assignment.services;
 
 import com.codesoom.assignment.domain.User;
 import com.codesoom.assignment.infra.JpaUserRepository;
