+++ app/src/test/java/com/codesoom/assignment/services/UserRegisterServiceTest.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.application;
+package com.codesoom.assignment.services;
 
 import com.codesoom.assignment.domain.User;
 import org.junit.jupiter.api.DisplayName;
